Working local copy hook and extraction refactoring
This is a pause mid refactor, to finish up I'd like to:
- ensure the metrics changes are brought back into the server
- hook up config for azure copy hook

This round of extraction refactor has been leaning on singletons (only
ones that are reasonable) in order to decouple / detangle some concepts.
Once this is done a second round of refactor should be doable in order
to move away from singletons to better dependency inversion.